moving config file to temp location before release
